+++ src/Data/Pool.hs
@@ -58,7 +58,7 @@ withResource pool act = mask $ \unmask -> do
 -- to the pool (via 'putResource').
 takeResource :: Pool a -> IO (a, LocalPool a)
 takeResource pool = mask_ $ do
-  lp <- getLocalPool (localPools pool)
+  lp <- getLocalPool (poolNumStripes $ poolConfig pool) (localPools pool)
   stripe <- takeMVar (stripeVar lp)
   if available stripe == 0
     then do
@@ -72,7 +72,7 @@ takeResource pool = mask_ $ do
     else takeAvailableResource pool lp stripe
 
 -- | A variant of 'withResource' that doesn't execute the action and returns
--- 'Nothing' instead of blocking if the capability-local pool is exhausted.
+-- 'Nothing' instead of blocking if the local pool is exhausted.
 tryWithResource :: Pool a -> (a -> IO r) -> IO (Maybe r)
 tryWithResource pool act = mask $ \unmask -> tryTakeResource pool >>= \case
   Just (res, localPool) -> do
@@ -82,10 +82,10 @@ tryWithResource pool act = mask $ \unmask -> tryTakeResource pool >>= \case
   Nothing -> pure Nothing
 
 -- | A variant of 'takeResource' that returns 'Nothing' instead of blocking if
--- the capability-local pool is exhausted.
+-- the local pool is exhausted.
 tryTakeResource :: Pool a -> IO (Maybe (a, LocalPool a))
 tryTakeResource pool = mask_ $ do
-  lp <- getLocalPool (localPools pool)
+  lp <- getLocalPool (poolNumStripes $ poolConfig pool) (localPools pool)
   stripe <- takeMVar (stripeVar lp)
   if available stripe == 0
     then do